Proofreading; demonic physics-enrichment programme
Gave this a thorough proof-reading for grammar (subject-verb and plural
agreement).  I also clarified and rearranged some sentences.  I
reinserted my idea about reaching physicists through submitting to ArXiv,
this time as a two-sentence comment.  This is a big problem in
evolutionary biology and it would be good to have more about it in
print.  I hope we can find some way to keep it.